requirements.txt
agno==1.7.5
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
beautifulsoup4==4.13.4
bidict==0.23.1
Brotli==1.1.0
bs4==0.0.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cssselect2==0.8.0
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.38.2
gradio_client==1.11.0
groovy==0.1.2
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
lancedb==0.24.1
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
minio==7.2.16
numpy==2.3.1
openai==1.97.1
orjson==3.11.1
overrides==7.7.0
packaging==25.0
panda==0.3.1
pandas==2.3.1
pillow==11.3.0
portalocker==3.2.0
protobuf==6.31.1
pyarrow==21.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pydyf==0.11.0
Pygments==2.19.2
pylance==0.31.1
pypdf==5.8.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.0
redis==6.2.0
requests==2.32.4
rich==14.0.0
ruff==0.12.5
safehttpx==0.1.6
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
tantivy==0.24.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unicorn==2.1.3
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
wsproto==1.2.0
youtube-search-python==1.6.6
youtube-transcript-api==0.6.2
zopfli==0.2.3.post1
